==> LIFE/scripts/05_MVMR_01_2Sample_SumStats_DataPrep.R <==
#' ---
#' title: "MVMR - 2-sample - summary statistics"
#' subtitle: "MR of SH on GE"
#' author: "Janne Pott"
#' date: "Last compiled on `r format(Sys.time(), '%d %B, %Y')`"
#' output:
#'   html_document:
#'     toc: true
#'     number_sections: true
#'     toc_float: true
#'     code_folding: show
#' ---
#'
#' # Introduction ####
#' ***
#' Plan: I want to do an MVMR of TT and BMI on GE. Here, I want to match the summary statistics data from TT and BMI, and then extract the GE summary statistics from the eQTLGen consortium. 
#' 
#' # Initialize ####
#' ***
rm(list = ls())
time0<-Sys.time()
server = "forostar"

source("../../SourceFile.R")
.libPaths()

source("../../helperfunctions/getSmallestDist.R")

#' # Load eQTLGen data ####
#' ***
eqtlgen = fread(path_eQTLGen)

all_genes = eqtlgen[,.N,by=c("Gene","GeneSymbol")]
all_SNPs = eqtlgen[,.N,by = c("SNP","SNPChr","SNPPos","AssessedAllele","OtherAllele")]

save(all_SNPs,file = "../temp/05_MVMR_eQTLGen_allSNPs.RData")
save(all_genes,file = "../temp/05_MVMR_eQTLGen_allgenes.RData")

#' # Get instruments ####
#' ***
myFiles = list.files(path = path_SumStats_QC)
myFiles = myFiles[!grepl("athway",myFiles)]
myFiles = myFiles[!grepl("SERPINA",myFiles)]
myFiles = myFiles[!grepl("pval",myFiles)]
myFiles = myFiles[!grepl("genomewide",myFiles)]
myFiles = myFiles[grepl("TT",myFiles)]

dumTab1 = foreach(i = 1:length(myFiles))%do%{
  #i=1
  
  # load TT data 
  erg2 = fread(paste0(path_SumStats_QC,myFiles[i]))
  erg2[,Zscore := BETA/SE]
  erg2[,Zscore_abs := abs(BETA/SE)]
  erg2 = erg2[rsID %in% all_SNPs$SNP,]
  
  # load BMI data
  if(i==1){
    erg3 = fread(path_Pulit_BMI_2019_men)
  }else{
    erg3 = fread(path_Pulit_BMI_2019_women)
  }
  erg3[,Zscore := BETA/SE]
  erg3[,Zscore_abs := abs(BETA/SE)]
  erg3[,rsID := gsub(":.*","",SNP)]
  erg3 = erg3[rsID %in% all_SNPs$SNP,]
  erg3 = erg3[!is.na(N),]
    
  # check overlap of TT and BMI and exclude triallelic SNPs
  erg2 = erg2[rsID %in% erg3$rsID]
  erg3 = erg3[rsID %in% erg2$rsID]
  
  dupSNPs = erg3[duplicated(rsID),rsID]
  erg3 = erg3[!is.element(rsID,dupSNPs)]
  erg2 = erg2[!is.element(rsID,dupSNPs)]
  matched = match(erg2$rsID,erg3$rsID)
  erg3 = erg3[matched,]
  
  # check alleles and harmonize (TT stays the same, BMI will be switched)
  stopifnot(erg2$rsID == erg3$rsID)
  erg2[,dumID := paste(rsID,OA,EA,sep="_")]
  erg3[,dumID1 := paste(rsID,Other_Allele,Tested_Allele,sep="_")]
  erg3[,dumID2 := paste(rsID,Tested_Allele,Other_Allele,sep="_")]
  erg3 = erg3[dumID1 %in% erg2$dumID | dumID2 %in% erg2$dumID,]
  erg2 = erg2[dumID %in% erg3$dumID1 | dumID %in% erg3$dumID2,]
  
  filt = erg2$EA == erg3$Tested_Allele
  erg3[,EA := Tested_Allele]
  erg3[,OA := Other_Allele]
  erg3[!filt,EA := Other_Allele]
  erg3[!filt,OA := Tested_Allele]
  erg3[!filt,BETA := BETA * (-1)]
  erg3[,EAF := Freq_Tested_Allele]
  erg3[!filt,EAF := 1-Freq_Tested_Allele]
  erg3[!filt,Zscore := Zscore * (-1)]
  erg3[,dumID3 := paste(rsID,OA,EA,sep="_")]
  stopifnot(erg3$dumID3 == erg2$dumID)
  
  # create dummy for pruning
  erg4 = copy(erg2)
  erg4[,Zscore_abs_BMI := erg3$Zscore_abs]
  erg4[,Zscore_max := pmax(Zscore_abs,Zscore_abs_BMI)]
  
  if(dim(erg4)[1]>0){
    # priority pruning
    setorder(erg4,CHR,BP)
    myCHRs = unique(erg4$CHR)
    
    result.22 = foreach(s2 = myCHRs) %do% {
      # s2 = myCHRs[1]
      subdata2 = copy(erg4)
      subdata2 = subdata2[CHR == s2, ]
      setkey(subdata2, BP)
      
      if(dim(subdata2)[1]<=1){
        subdata2[, keep := T]
        subdata2[, NR_SNPs := 0]
      }else{
        subdata2[, keep := NA]
        subdata2[, NR_SNPs := as.numeric(NA)]
        
        smallestDist = getSmallestDist(subdata2[, BP])
        while(smallestDist < 1000000) {
          #minP = min(subdata2[is.na(keep), p])
          #maxLogP = max(subdata2[is.na(keep), pvalue_neg_log10_GC])
          maxAbsZscore = max(subdata2[is.na(keep), Zscore_max])
          myPOS = subdata2[maxAbsZscore == Zscore_max & is.na(keep), BP]
          if(length(myPOS)>1){
            myPOS = myPOS[1]
          }
          subdata2[BP == myPOS, keep := T]
          
          #filter for SNPs that can stay within the set (outside the +- 1MB range or keep==T)
          myFilt = (subdata2[, BP] < (myPOS - 1000000)) | 
            (subdata2[, BP] > (myPOS + 1000000)) | 
            subdata2[, keep] 
          myFilt[is.na(myFilt)] = FALSE
          subdata2 = subdata2[myFilt == TRUE, ]
          
          subdata2[BP == myPOS, NR_SNPs := sum(myFilt==F)]
          smallestDist = getSmallestDist(subdata2[, BP])
        }
        
        #stopifnot(sum(is.na(subdata2[,keep])) <= 1)
        subdata2[is.na(keep), NR_SNPs := 0]
        subdata2[is.na(keep), keep := TRUE]
      }
      
      subdata2
    }
    erg5 = rbindlist(result.22)
    myFile2 = gsub(".txt.gz","",myFiles[i])
    outfile = paste0("../temp/05_MVMR_BMI_",myFile2,"_genomewide_eqtlgenMatched.txt")
    fwrite(erg5, file = outfile, quote = F, sep = "\t", col.names = T, row.names = F, na = NA, dec = ".")
  }
  
  # filter for independent and significant SNPs
  erg5 = erg5[Zscore_max>5,]
  erg2 = erg2[rsID %in% erg5$rsID,]
  erg3 = erg3[rsID %in% erg5$rsID,]

  erg6 = cbind(erg2[,c(1,14,17,13,2:5,9,8,6,15,7,10)],erg3[,c(19,7,8,12,9,10)])
  names(erg6)[9:14] = paste(names(erg6)[9:14],"TT",sep = "_")
  names(erg6)[15:20] = paste(names(erg6)[15:20],"BMI",sep = "_")
  erg6
}
erg = rbindlist(dumTab1,use.names = T,fill=T)
save(erg,file = "../temp/05_MVMR_BMI_TT_IVs.RData")

table(is.element(erg$rsID,all_SNPs$SNP))

#' # Prep eQTL data ####
#' ***
#' Restrict to instruments, check alleles, get LIFE-Adult EAF, calculate beta and SE 
#' 
filtered_SNPs = copy(all_SNPs)
filtered_SNPs = filtered_SNPs[SNP %in% erg$rsID,]

load("../results/02_PGS_01_LIFE_pvar_filtered.RData")
table(is.element(erg$ID,pvar$ID))
pvar = pvar[ID %in% erg$ID,]
matched1 = match(pvar$ID,erg$ID)
stopifnot(pvar$ID == erg$ID[matched1])
pvar[,rsID := erg[matched1,rsID]]
stopifnot(pvar$rsID %in% filtered_SNPs$SNP)

matched2 = match(pvar$rsID,filtered_SNPs$SNP)
filtered_SNPs = filtered_SNPs[matched2,]

stopifnot(filtered_SNPs$SNP == pvar$rsID)
stopifnot(filtered_SNPs$SNPChr == pvar$CHR)
plot(filtered_SNPs$SNPPos, pvar$POS)

table(filtered_SNPs$AssessedAllele == pvar$ALT,filtered_SNPs$OtherAllele == pvar$REF)
table(filtered_SNPs$OtherAllele == pvar$ALT,filtered_SNPs$AssessedAllele == pvar$REF)
table(erg[matched1,EA] == pvar$ALT,erg[matched1,OA] == pvar$REF)
filtered_SNPs[,switchAlleles := F]
filtered_SNPs[OtherAllele == pvar$ALT, switchAlleles := T]
filtered_SNPs[,EAF_LIFE := pvar$AF_Adult]
filtered_SNPs[,MAF_LIFE := EAF_LIFE]
filtered_SNPs[EAF_LIFE>0.5,MAF_LIFE := 1-EAF_LIFE]
plot(filtered_SNPs$EAF_LIFE,filtered_SNPs$MAF_LIFE)

eqtlgen_filtered = eqtlgen[SNP %in% filtered_SNPs$SNP,]

dumTab2 = foreach(i = 1:dim(filtered_SNPs)[1])%do%{
  #i=1
  mySNP = filtered_SNPs[i,]
  erg5 = copy(eqtlgen_filtered)
  erg5 = erg5[SNP == mySNP$SNP,]
  
  if(mySNP$switchAlleles==T){
    erg5[,EA := OtherAllele]
    erg5[,OA := AssessedAllele]
    erg5[,Zscore_harmonized := Zscore * (-1)]
  }else{
    erg5[,EA := AssessedAllele]
    erg5[,OA := OtherAllele]
    erg5[,Zscore_harmonized := Zscore]
  }
  
  erg5[, beta := Zscore_harmonized / sqrt(NrSamples * 2 * mySNP$MAF_LIFE * (1-mySNP$MAF_LIFE))]
  erg5[, SE := beta/Zscore_harmonized]
  erg5 = erg5[beta!=0,]
  erg5
}
eqtlgen_harmonized = rbindlist(dumTab2)
save(eqtlgen_harmonized,file = "../temp/05_MVMR_eqtlgen_harmonized.RData")

#' # Session Info ####
#' ***
sessionInfo()
message("\nTOTAL TIME : " ,round(difftime(Sys.time(),time0,units = "mins"),3)," minutes")
